# docs/swagger.yaml
definitions:
...
  handlers.CreateBlogPostRequest:
    properties:
      category:
        type: string
      content:
        type: string
      excerpt:
        type: string
      is_featured:
        type: boolean
      is_published:
        type: boolean
      localizations:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        description: Map of language code to localized fields
        type: object
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - category
    - content
    - title
    type: object
...
info:
  contact: {}
paths:
...
  /admin/blog:
    post:
      consumes:
      - application/json
      description: Creates a new blog post (admin only)
      parameters:
      - description: Blog post information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBlogPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Blog post created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only admins can create blog posts
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new blog post
      tags:
      - blog
      - admin
  /admin/blog/{post_id}:
    delete:
      description: Deletes an existing blog post (admin only)
      parameters:
      - description: Blog post ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog post deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only admins can delete blog posts
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Blog post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a blog post
      tags:
      - blog
      - admin
    put:
      consumes:
      - application/json
      description: Updates an existing blog post (admin only)
      parameters:
      - description: Blog post ID
        in: path
        name: post_id
        required: true
        type: integer
      - description: Updated blog post information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBlogPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Blog post updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only admins can update blog posts
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Blog post not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a blog post
      tags:
      - blog
      - admin
...
  /blog:
    get:
      description: Retrieves all published blog posts with optional filtering
      parameters:
      - description: Filter by category
        in: query
        name: category
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Language for localized content
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of blog posts
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all published blog posts
      tags:
      - blog
  /blog/{slug}:
    get:
      description: Retrieves a specific blog post by its slug
      parameters:
      - description: Blog post slug
        in: path
        name: slug
        required: true
        type: string
      - description: Language for localized content
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Blog post details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Blog post not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific blog post
      tags:
      - blog
  /blog/categories:
    get:
      description: Retrieves all distinct categories from published blog posts
      produces:
      - application/json
      responses:
        "200":
          description: List of blog categories
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all blog categories
      tags:
      - blog
  /blog/featured:
    get:
      description: Retrieves all featured and published blog posts
      parameters:
      - description: Language for localized content
        in: query
        name: language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of featured blog posts
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get featured blog posts
      tags:
      - blog
  /blog/tags:
    get:
      description: Retrieves all unique tags from published blog posts
      produces:
      - application/json
      responses:
        "200":
          description: List of blog tags
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all blog tags
      tags:
      - blog
...
swagger: "2.0"